Address PR review: move lock to src/platform, unify temp test helper
- Move cross_process_file_lock.{h,cc} from src/download into src/platform
  so the generic orchestration lives alongside its per-OS implementations
  (matching the ORT layout), instead of split across two directories.
  Update includes, CMakeLists source path, comments, and the plan doc.
- Combine the TempDir and TempPath test helpers into a single TempPath
  (test/utils/temp_path.h) with a private constructor and CreateTempDir /
  CreateTempFile factory methods.
- Guard the TempPath destructor with try/catch and a stderr diagnostic so a
  cleanup failure can never escape the destructor.
