+++ sdk_v2/cpp/src/platform/cross_process_file_lock.h
