+++ sdk_v2/cpp/src/platform/windows/cross_process_file_lock.cc
@@ -2,8 +2,8 @@
 // Licensed under the MIT License.
 // Windows implementation of CrossProcessFileLock acquisition and the
 // platform-specific lock handle. Cross-platform orchestration
-// (WaitForDirectoryLock) lives in src/download/cross_process_file_lock.cc.
-#include "download/cross_process_file_lock.h"
+// (WaitForDirectoryLock) lives in src/platform/cross_process_file_lock.cc.
+#include "platform/cross_process_file_lock.h"
 #include "exception.h"
 #include "logger.h"
